Chapters 14.5 and 14.6
This adds more complex logic to the VM, so that it now has the ability
to handle constant expressions as well.

I've had to deviate slightly from the way that the C version was
implemented, in order to be more Rust idiomatic. We can be nice and
succinct by ensuring that the instructions are all one byte long rather
than requiring varying length.